refactor: bio 타입 위치 수정 및 임포트 수정
