+++ src/features/bio/index.ts
@@ -1,1 +1,2 @@
 export { default as BioRoot } from './components/BioRoot';
+export { bioFormSchema, type UpdateBioType } from './types/bio.types';
